+++ src/Model/Model.Common/PlayerStatistic.cs
@@ -2,7 +2,7 @@ namespace App.Models;
 
 public class PlayerStatistic
 {
-	public int AverageTurnScore { get; set; }
+	public double AverageTurnScore { get; set; }
 	public int PlayedTurns { get; set; }
 	public int OneEighties { get; set; }
 	public int Ninedarters { get; set; }
@@ -15,7 +15,7 @@ public class PlayerStatistic
 		this.Ninedarters = 0;
 	}
 
-	public PlayerStatistic(int averageTurnScore, int playedTurns, int oneEighties, int ninedarters)
+	public PlayerStatistic(double averageTurnScore, int playedTurns, int oneEighties, int ninedarters)
 	{
 		this.AverageTurnScore = averageTurnScore;
 		this.PlayedTurns = playedTurns;
